nuovo file:             src/currenttitle/option/main_exe.cpp 	nuovo file:             src/currenttitle/option/ui_option_copybook.cpp 	nuovo file:             src/currenttitle/option/ui_option_copybook.h 	nuovo file:             src/currenttitle/option/ui_option_copybook.ui 	modificato:             src/mainwindow.cpp 	modificato:             src/mainwindow.h 	modificato:             src/mainwindow.ui 	modificato:             src/touch/pageCount.cpp 	modificato:             src/touch/pen/pen_ui.h 	modificato:             src/touch/tabletevent.cpp 	modificato:             writernote.pro

+++ src/touch/pen/pen_ui.h
@@ -62,7 +62,7 @@ public:
         }
     }
 
-    void setWidthTratto(const int size){
+    void setWidthTratto(const double size){
         m_spessore_pen = size;
         list_update();
     }
